move xdist hook
