+++ ddtrace/contrib/internal/pytest/plugin.py
@@ -180,29 +180,3 @@ def patch_all(request):
         from ddtrace._monkey import _patch_all
 
         _patch_all()
-
-
-@pytest.hookimpl
-def pytest_configure_node(node):
-    # compute or read whatever you need on master
-    from ddtrace.internal.logger import get_logger
-    from ddtrace.internal.test_visibility.api import InternalTestSession
-
-    log = get_logger(__name__)
-    main_session_span = InternalTestSession.get_span()
-    if main_session_span:
-        root_trace = main_session_span.trace_id
-        root_span = main_session_span.span_id
-        log.debug(
-            "pytest_configure_node (main): Main session span_id: %s, trace_id: %s. Passing to worker.",
-            root_span,
-            root_trace,
-        )
-    else:
-        log.error("pytest_configure_node (main): Could not get main session span!")
-        root_trace = 0
-        root_span = 0
-
-    log.debug("Setting root_span %s and root_trace %s", root_span, root_trace)
-    node.workerinput["root_span"] = root_span
-    node.workerinput["root_trace"] = root_trace
